Crate missing

The Missing Semester of Your CS Education

https://missing.csail.mit.edu

https://www.youtube.com/playlist?list=PLyzOVJj3bHQunmnnTXrNbZnBaCA-ieK4L

https://news.ycombinator.com/item?id=47124171

How this doc is generated

missing/Makefile

all:	doc not-found

doc:
	rustdoc missing.rs -o ../docs
	rustdoc not-found.rs -o ../docs
	rustdoc vi-wrap.rs -o ../docs
	rustdoc fmt-wrap.rs -o ../docs
	rustdoc get-env.rs -o ../docs

not-found:
	rustfmt not-found.rs
	rustc not-found.rs
	-./not-found 1> not-found.stdout 2> not-found.stderr

get-env:
	rustfmt get-env.rs
	rustc get-env.rs

clean:
	rm -f env

You can run following shell command from the root level

make -C missing

RTFM:

make --help
Usage: make [options] [target] ...
Options:
...
 -C DIRECTORY, --directory=DIRECTORY
                             Change to DIRECTORY before doing anything.

Why missing

When I was a Physics Ph.D. student in NYU in the late nineties, I took a course called UNIX tools in the CS department. It was a hands on course where the instructor did live REPL in the terminal and we watched him showing us all the tricks. I got hooked with UNIX since then. Got myself a dialup terminal in my tiny apartment in east village and dial in to the workstation on campus. The latency is so bad that I can’t only see the feedback after a few keystrokes. That was when I trained my vi muscle memory. (EMacs was out of the question.)

Later I got my own IBM 386 and installed Linux on it and started to program in Perl …

I am a big fan of Jon’s YouTube videos on Rust and I started to use Rust in non conventional ways.

I am going to follow this lecture series and “port” them to rustdoc and see how it goes.

What was missing then

// TODO

What is missing now

// TODO